# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Outfitte API
  version: "5"
  description: |
    REST API for Outfitte — a self-hosted wardrobe management application.
    Current milestone: M5 (Granular Sharing).
servers:
  - url: http://localhost:8080
    description: Local development
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          example: not found
...
    Photo:
      type: object
      required: [id, media_key, position, created_at]
      properties:
        id:
          type: string
          format: uuid
        media_key:
          type: string
        position:
          type: integer
        created_at:
          type: string
          format: date-time
    Item:
      type: object
      required: [id, owner_id, name, brand, category_id, color, metadata, photos, location_id, purchase_price, purchase_currency, purchase_date, seller_url, created_at]
      properties:
        id:
          type: string
          format: uuid
        owner_id:
          type: string
          format: uuid
        name:
          type: string
        brand:
          type: [string, "null"]
        category_id:
          type: [string, "null"]
          format: uuid
        color:
          type: [string, "null"]
        metadata:
          type: object
          additionalProperties:
            type: string
        photos:
          type: array
          items:
            $ref: '#/components/schemas/Photo'
        location_id:
          type: [string, "null"]
          format: uuid
        purchase_price:
          type: [string, "null"]
          description: Decimal string, e.g. "29.99". Must be accompanied by purchase_currency (both set or both absent).
        purchase_currency:
          type: [string, "null"]
          description: 3-letter ISO 4217 currency code (case-insensitive input, stored uppercase). Must be accompanied by purchase_price.
          example: "USD"
        purchase_date:
          type: [string, "null"]
          format: date
          description: Purchase date in YYYY-MM-DD format. Must not be a future date.
          example: "2023-06-15"
        seller_url:
          type: [string, "null"]
          description: URL of the seller or product page.
          example: "https://example.com/product"
        created_at:
          type: string
          format: date-time
...
    OutfitItem:
      type: object
      required: [outfit_id, item_id, position]
      properties:
        outfit_id:
          type: string
          format: uuid
        item_id:
          type: string
          format: uuid
        position:
          type: integer
    Outfit:
      type: object
      required: [id, owner_id, name, notes, items, photos, created_at]
      properties:
        id:
          type: string
          format: uuid
        owner_id:
          type: string
          format: uuid
        name:
          type: [string, "null"]
        notes:
          type: [string, "null"]
        items:
          type: array
          items:
            $ref: '#/components/schemas/OutfitItem'
        photos:
          type: array
          items:
            $ref: '#/components/schemas/Photo'
        created_at:
          type: string
          format: date-time
...
    Location:
      type: object
      required: [id, owner_id, parent_id, label, created_at]
      properties:
        id:
          type: string
          format: uuid
        owner_id:
          type: string
          format: uuid
        parent_id:
          type: [string, "null"]
          format: uuid
        label:
          type: string
        created_at:
          type: string
          format: date-time
...
    UserSummary:
      type: object
      required: [id, email]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
...
    SharedItem:
      allOf:
        - $ref: '#/components/schemas/Item'
        - type: object
          required: [shared_by]
          properties:
            shared_by:
              $ref: '#/components/schemas/UserSummary'
    SharedOutfit:
      allOf:
        - $ref: '#/components/schemas/Outfit'
        - type: object
          required: [shared_by]
          properties:
            shared_by:
              $ref: '#/components/schemas/UserSummary'
    SharedLocation:
      type: object
      required: [location, items, shared_by]
      properties:
        location:
          $ref: '#/components/schemas/Location'
        items:
          type: array
          items:
            $ref: '#/components/schemas/Item'
          description: Items in the shared location (location inheritance — items inherit access from their location)
        shared_by:
          $ref: '#/components/schemas/UserSummary'
    SharedWithMeResult:
      type: object
      required: [items, outfits, locations]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/SharedItem'
        outfits:
          type: array
          items:
            $ref: '#/components/schemas/SharedOutfit'
        locations:
          type: array
          items:
            $ref: '#/components/schemas/SharedLocation'
  responses:
...
    Unauthorized:
      description: Missing or invalid JWT
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: unauthorized
...
security: []
paths:
...
  /shares/with-me:
    get:
      tags: [shares]
      summary: List entities shared with me
      operationId: listSharedWithMe
      security:
        - bearerAuth: []
      responses:
        "200":
          description: All entities shared with the caller, grouped by type
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SharedWithMeResult'
        "401":
          $ref: '#/components/responses/Unauthorized'
...